# models/index.js
const Exercises = require('./exercises');
const ExercisesDiary = require('./exercisesInDiary');
const ExerciseSubCategories = require('./exerciseSubCategories');
const Products = require('./products');
const ProductsCategories = require('./productsCategories');
const ProductsDiary = require('./productsInDiary');
const User = require('./user');
const Review = require('./review');

module.exports = {
  Exercises,
  ExercisesDiary,
  ExerciseSubCategories,
  Products,
  ProductsCategories,
  ProductsDiary,
  User,
  Review,
};