# swagger.json
{
  "openapi": "3.0.2",
  "info": {
    "title": "Power Pulse API",
    "version": "1.0",
    "description": "Demonstrate Power Pulse API"
  },
  "servers": [
    {
      "url": "https://power-pulse-api.onrender.com/api",
      "description": "Server on render.com"
    },
    {
      "url": "http://localhost:3000/api",
      "description": "Local server"
    }
  ],
  "components": {
    "schemas": {
      "Name": {
        "type": "string",
        "description": "User's name",
        "example": "David"
      },
...
      "AvatarUrls": {
        "type": "object",
        "description": "User's avatar",
        "properties": {
          "avatar_37x37": {
            "type": "string",
            "example": "https://res.cloudinary.com/dhz4yf0av/image/upload/h_37,w_37/power_pulse_avatars/avatar_65356d56fa7a5576893e8407.jpg"
          },
          "avatar_46x46": {
            "type": "string",
            "example": "https://res.cloudinary.com/dhz4yf0av/image/upload/h_46,w_46/power_pulse_avatars/avatar_65356d56fa7a5576893e8407.jpg"
          },
          "avatar_74x74": {
            "type": "string",
            "example": "https://res.cloudinary.com/dhz4yf0av/image/upload/h_74,w_74/power_pulse_avatars/avatar_65356d56fa7a5576893e8407.jpg"
          },
          "avatar_90x90": {
            "type": "string",
            "example": "https://res.cloudinary.com/dhz4yf0av/image/upload/h_90,w_90/power_pulse_avatars/avatar_65356d56fa7a5576893e8407.jpg"
          },
          "avatar_92x92": {
            "type": "string",
            "example": "https://res.cloudinary.com/dhz4yf0av/image/upload/h_92,w_92/power_pulse_avatars/avatar_65356d56fa7a5576893e8407.jpg"
          },
          "avatar_140x140": {
            "type": "string",
            "example": "https://res.cloudinary.com/dhz4yf0av/image/upload/h_140,w_140/power_pulse_avatars/avatar_65356d56fa7a5576893e8407.jpg"
          },
          "avatar_150x150": {
            "type": "string",
            "example": "https://res.cloudinary.com/dhz4yf0av/image/upload/h_150,w_150/power_pulse_avatars/avatar_65356d56fa7a5576893e8407.jpg"
          },
          "avatar_180x180": {
            "type": "string",
            "example": "https://res.cloudinary.com/dhz4yf0av/image/upload/h_180,w_180/power_pulse_avatars/avatar_65356d56fa7a5576893e8407.jpg"
          },
          "avatar_250x250": {
            "type": "string",
            "example": "https://res.cloudinary.com/dhz4yf0av/image/upload/h_250,w_250/power_pulse_avatars/avatar_65356d56fa7a5576893e8407.jpg"
          },
          "avatar_300x300": {
            "type": "string",
            "example": "https://res.cloudinary.com/dhz4yf0av/image/upload/h_300,w_300/power_pulse_avatars/avatar_65356d56fa7a5576893e8407.jpg"
          },
          "avatar_350x350": {
            "type": "string",
            "example": "https://res.cloudinary.com/dhz4yf0av/image/upload/h_350,w_350/power_pulse_avatars/avatar_65356d56fa7a5576893e8407.jpg"
          },
          "avatar_450x450": {
            "type": "string",
            "example": "https://res.cloudinary.com/dhz4yf0av/image/upload/h_450,w_450/power_pulse_avatars/avatar_65356d56fa7a5576893e8407.jpg"
          }
        }
      },
      "CreatedAt": {
        "type": "string",
        "description": "Time when user was created at",
        "format": "Time",
        "example": "2023-10-22T18:43:34.175Z"
      },
...
      "_Id": {
        "type": "string",
        "example": "6540e173fcec8c3a63bbd7b4"
      },
...
      "Rating": {
        "type": "number",
        "description": "Review's rating",
        "example": 5
      },
      "Description": {
        "type": "string",
        "description": "Review's description",
        "example": "Very good"
      },
...
      "Review": {
        "type": "object",
        "properties": {
          "_id": {
            "$ref": "#/components/schemas/_Id"
          },
          "rating": {
            "$ref": "#/components/schemas/Rating"
          },
          "description": {
            "$ref": "#/components/schemas/Description"
          },
          "owner": {
            "type": "object",
            "properties": {
              "_id": {
                "$ref": "#/components/schemas/_Id"
              },
              "name": {
                "$ref": "#/components/schemas/Name"
              },
              "avatarUrls": {
                "$ref": "#/components/schemas/AvatarUrls"
              }
            }
          },
          "createdAt": {
            "$ref": "#/components/schemas/CreatedAt"
          },
          "updatedAt": {
            "$ref": "#/components/schemas/CreatedAt"
          }
        }
      },
...
      "UpdateReviewRequest": {
        "type": "object",
        "properties": {
          "rating": {
            "$ref": "#/components/schemas/Rating"
          },
          "description": {
            "$ref": "#/components/schemas/Description"
          }
        }
      },
...
      "ErrorMessageResponse": {
        "type": "object",
        "properties": {
          "message": {
            "type": "string"
          }
        }
      }
    },
    "securitySchemes": {
      "Bearer": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    }
  },
...
  "paths": {
...
    "/reviews/private/{id}": {
      "get": {
        "tags": ["Reviews"],
        "summary": "Return user's review by id",
        "operationId": "getPrivateReviewById",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "Choose user's review by id",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "security": [{ "Bearer": [] }],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Review"
                }
              }
            }
          },
          "401": {
            "description": "Authorization error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorMessageResponse"
                }
              }
            }
          },
          "404": {
            "description": "Not Found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorMessageResponse"
                }
              }
            }
          }
        }
      },
      "patch": {
        "tags": ["Reviews"],
        "summary": "Update user's review by id",
        "operationId": "updatePrivateReviewById",
        "security": [
          {
            "Bearer": []
          }
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "Choose user's review by id",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "description": "Review's's object",
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateReviewRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Review"
                }
              }
            }
          },
          "401": {
            "description": "Authorization error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorMessageResponse"
                }
              }
            }
          },
          "404": {
            "description": "Not Found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorMessageResponse"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": ["Reviews"],
        "summary": "Delete user's review by id",
        "operationId": "removePrivateReviewById",
        "security": [
          {
            "Bearer": []
          }
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "Choose review by id",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Review"
                }
              }
            }
          },
          "401": {
            "description": "Authorization error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorMessageResponse"
                }
              }
            }
          },
          "404": {
            "description": "Not Found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorMessageResponse"
                }
              }
            }
          }
        }
      }
    }
  }
}